Add a separate module for pre-translation
