+++ pontoon/base/models.py
@@ -56,6 +56,7 @@ from pontoon.sync.vcs.repositories import (
     update_from_vcs,
     PullFromRepositoryException,
 )
+from pontoon.pretranslation.utils import pre_translate
 
 
 log = logging.getLogger(__name__)
@@ -2684,7 +2685,7 @@ class Translation(DirtyFieldsMixin, models.Model):
 
     # Active translations are displayed in the string list and as the first
     # entry in the History tab. There can only be one active translation for
-    # each (entity, locale, plural_foclass rm) combination. See bug 1481175.
+    # each (entity, locale, plural_form) combination. See bug 1481175.
     active = models.BooleanField(default=False)
 
     fuzzy = models.BooleanField(default=False)
@@ -3355,73 +3356,3 @@ class TranslatedResource(AggregatedStats):
             strings_with_warnings_diff,
             unreviewed_strings_diff,
         )
-
-
-from pontoon.machinery.views import translation_memory
-from pontoon.machinery.views import google_translate
-
-
-def pre_translate(project, locales=None, entities=None):
-    """
-    Identifies strings without any translations and any suggestions.
-    Engages TheAlgorithm (bug 1552796) to gather pre-translations.
-    Stores pre-translations as suggestions (approved=False) to DB.
-    Author should be set to a name like "Pontoon Translator <pontoon@mozilla.com>".
-    Stats and Latest activity should be updated
-
-    """
-
-    if not locales:
-        locales = project.locales.filter(project_locale__readonly=False)
-
-    if not entities:
-        entities = Entity.objects.filter(
-            resource__project=project,
-            obsolete=False,
-            translation__isnull=True,
-        )
-
-    now = timezone.now()
-
-    resources = project.resources.filter(obsolete=False)
-
-    user = User.objects.get(pk=2)
-    for locale in locales:
-        for entity in entities:
-            string = ""
-            tm_response = translation_memory(
-                text=entity.string,
-                locale=locale.code,
-            )
-
-            tm_response = json.loads(tm_response.content)
-
-            if tm_response and int(float(tm_response[0]['quality'])) is 100:
-                string = tm_response[0]['target']
-
-            elif locale.google_translate_code:
-                gt_response = google_translate(
-                    text=entity.string,
-                    locale_code=locale.google_translate_code,
-                )
-
-                gt_response = json.loads(gt_response.content)
-
-                if 'translation' in gt_response.keys():
-                    string = gt_response['translation']
-
-            if string:
-                t = Translation(
-                    entity=entity,
-                    locale=locale,
-                    string=string,
-                    user=user,
-                    date=now,
-                    approved=False,
-                )
-                t.save()
-                t.update_latest_translation()
-
-        for resource in resources:
-            translatedresource = TranslatedResource.objects.get(resource=resource, locale=locale)
-            translatedresource.calculate_stats()
